Creacion de nuevos servicios y correcion de anteriores

+++ src/main/java/com/example/demo/dao/CarritoDAO.java
@@ -3,4 +3,4 @@ package com.example.demo.dao;
 import com.example.demo.model.Carrito;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CarritoDAO extends JpaRepository<Carrito, Integer>{}
+public interface CarritoDAO extends JpaRepository<Carrito, Long>{}
